+++ .cargo/config.toml
@@ -1,16 +1,12 @@
-[unstable]
-build-std = ["core", "compiler_builtins", "alloc"]
-build-std-features = ["compiler-builtins-mem"]
-
 [build]
 # https://internals.rust-lang.org/t/evaluating-pipelined-rustc-compilation/10199/12
 pipelining = true
 
 [target.aarch64-vesper-metta]
 rustflags = [
-  "-C", "target-feature=-fp-armv8",
-  "-C", "target-cpu=cortex-a53",
-  "-C", "embed-bitcode=yes",
-  "-Z", "macro-backtrace",
+    "-C", "target-feature=-fp-armv8",
+    "-C", "target-cpu=cortex-a53",
+    "-C", "embed-bitcode=yes",
+    "-Z", "macro-backtrace",
 ]
 runner = "cargo make test-runner"
