+++ machine/src/macros.rs
