+++ machine/src/platform/rpi3/vc.rs
