+++ machine/src/write_to.rs
