+++ nucleus/Makefile.toml
@@ -3,92 +3,28 @@
 #
 # Copyright (c) Berkus Decker <berkus+vesper@metta.systems>
 #
-[tasks.build]
-env = { "TARGET_FEATURES" = "" }
-args = ["build", "--target=${TARGET_JSON}", "--release", "--features=${TARGET_FEATURES}"]
-
-[tasks.expand]
-env = { "TARGET_FEATURES" = "" }
-args = ["expand", "--target=${TARGET_JSON}", "--release", "--features=${TARGET_FEATURES}"]
-
-[tasks.test]
-env = { "TARGET_FEATURES" = "${QEMU_FEATURES}" }
-args = ["test", "--target=${TARGET_JSON}", "--features=${TARGET_FEATURES}"]
-
-[tasks.docs]
-env = { "TARGET_FEATURES" = "" }
-args = ["doc", "--open", "--no-deps", "--target=${TARGET_JSON}", "--features=${TARGET_FEATURES}"]
-
-# These tasks are written in cargo-make's own script to make it portable across platforms (no `basename` on Windows)
 [tasks.kernel-binary]
-script_runner = "@duckscript"
-script = [
-'''
-    cp ${CARGO_MAKE_WORKSPACE_WORKING_DIRECTORY}/target/${TARGET}/release/vesper ${KERNEL_ELF}
-    exec --fail-on-error ${OBJCOPY} %{OBJCOPY_PARAMS} ${CARGO_MAKE_WORKSPACE_WORKING_DIRECTORY}/target/${TARGET}/release/vesper ${KERNEL_BIN}
-'''
-]
-
-[tasks.custom-binary]
-script_runner = "@duckscript"
-script = [
-'''
-    binaryFile = basename ${CARGO_MAKE_TASK_ARGS}
-    cp ${CARGO_MAKE_TASK_ARGS} ${CARGO_MAKE_WORKSPACE_WORKING_DIRECTORY}/target/${binaryFile}.elf
-    exec --fail-on-error ${OBJCOPY} %{OBJCOPY_PARAMS} ${CARGO_MAKE_TASK_ARGS} ${CARGO_MAKE_WORKSPACE_WORKING_DIRECTORY}/target/${binaryFile}.bin
-'''
-]
-
-[tasks.test-runner]
-dependencies = ["custom-binary"]
-script_runner = "@duckscript"
-script = [
-'''
-    binaryFile = basename ${CARGO_MAKE_TASK_ARGS}
-    exec --fail-on-error ${QEMU} %{QEMU_OPTS} %{QEMU_TESTS_OPTS} -dtb ${CARGO_MAKE_WORKSPACE_WORKING_DIRECTORY}/targets/bcm2710-rpi-3-b-plus.dtb -kernel ${CARGO_MAKE_WORKSPACE_WORKING_DIRECTORY}/target/${binaryFile}.bin
-'''
-]
-
-[tasks.build-qemu]
-env = { "TARGET_FEATURES" = "${QEMU_FEATURES}" }
-command = "cargo"
-args = ["build", "--target=${TARGET_JSON}", "--release", "--features=${TARGET_FEATURES}"]
+env = { "BINARY_FILE" = "${KERNEL_ELF}" }
+run_task = "custom-binary"
 
 [tasks.qemu-runner]
+dependencies = ["build-qemu", "kernel-binary"]
 env = { "TARGET_FEATURES" = "${QEMU_FEATURES}" }
 script = [
     "${QEMU} ${QEMU_OPTS} ${QEMU_RUNNER_OPTS} -dtb ${CARGO_MAKE_WORKSPACE_WORKING_DIRECTORY}/targets/bcm2710-rpi-3-b-plus.dtb -kernel ${KERNEL_BIN}"
 ]
-dependencies = ["build-qemu", "kernel-binary"]
 
 [tasks.qemu]
 extend = "qemu-runner"
 env = { "QEMU_RUNNER_OPTS" = "${QEMU_SERIAL_OPTS}" }
 
-[tasks.zellij-config]
-extend = "qemu-runner"
-script_runner = "@duckscript"
-env = { "ZELLIJ_CONFIG_FILE" = "../emulation/zellij-config.sh" }
-script = [
-'''
-    writefile ${ZELLIJ_CONFIG_FILE} "QEMU=${QEMU}\n"
-    appendfile ${ZELLIJ_CONFIG_FILE} "QEMU_OPTS=\"${QEMU_OPTS}\"\n"
-    appendfile ${ZELLIJ_CONFIG_FILE} "QEMU_RUNNER_OPTS=${QEMU_RUNNER_OPTS}\n"
-    appendfile ${ZELLIJ_CONFIG_FILE} "CARGO_MAKE_WORKSPACE_WORKING_DIRECTORY=${CARGO_MAKE_WORKSPACE_WORKING_DIRECTORY}\n"
-    appendfile ${ZELLIJ_CONFIG_FILE} "KERNEL_BIN=${KERNEL_BIN}\n"
-'''
-]
-install_crate = { crate_name = "zellij", binary = "zellij", test_arg = ["--help"] }
-
 [tasks.qemu-gdb]
 extend = "qemu-runner"
 env = { "QEMU_RUNNER_OPTS" = "${QEMU_SERIAL_OPTS} ${QEMU_GDB_OPTS}" }
 
-[tasks.openocd]
-dependencies = ["build", "kernel-binary"]
-script = [
-    "${OPENOCD} -f interface/jlink.cfg -f ../ocd/${TARGET_BOARD}_target.cfg"
-]
+[tasks.zellij-nucleus]
+env = { "KERNEL_BIN" = "${KERNEL_BIN}" }
+run_task = "zellij-config"
 
 [tasks.gdb-config]
 script_runner = "@duckscript"
@@ -110,7 +46,7 @@ dependencies = ["build", "kernel-binary"]
 script = [
     "${NM} -- ${KERNEL_ELF} | sort"
 ]
-#install_crate = "cargo-binutils"
+install_crate = { crate_name = "cargo-binutils", binary = "rust-nm", test_arg = ["--help"] }
 
 [tasks.sdcard]
 dependencies = ["build", "kernel-binary"]
@@ -121,11 +57,6 @@ script = [
     cp ${KERNEL_BIN} ${VOLUME}/${kernelImage}
 '''
 ]
-
-[tasks.clippy]
-env = { "TARGET_FEATURES" = { value = "--features=${CLIPPY_FEATURES}", condition = { env_set = ["CLIPPY_FEATURES"] } } }
-command = "cargo"
-args = ["clippy", "--target=${TARGET_JSON}", "@@remove-empty(TARGET_FEATURES)", "--", "-D", "warnings"]
 
 [tasks.hopper]
 dependencies = ["build", "kernel-binary"]
